% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/describe.R
\name{category}
\alias{category}
\title{Return list of HarvestChoice variable categories}
\usage{
category(cat)
}
\arguments{
\item{cat}{optional category filter}
}
\value{
a data.table of variable categories
}
\description{
Return list of HarvestChoice variable categories
}

