% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{cell5m-meta}
\alias{cell5m-meta}
\title{HarvestChoice CELL5M Variable Inventory}
\format{data.table objects}
\usage{
data(cell5m-meta)
}
\description{
Metadata for HarvestChoice collection of over 700 5-arc-minute rasters for sub-Saharan Africa.
}
\keyword{datasets}

