% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapackage.R
\name{datapackage}
\alias{datapackage}
\title{Write auxiliary files to CSV and JSON (Tabular Data Package RFC specifications)}
\usage{
datapackage(var, format = NULL, dir = ".")
}
\arguments{
\item{var}{character array of HarvestChoice indicator code(s)}

\item{format}{file extension corresponding to the data package output
format. See \code{\link{genFile}}. If \code{NULL} no auxiliary file is produced
aside from metadata records in \code{meta.csv}.}

\item{dir}{character, optional output directory. Default to current working directory.}
}
\value{
character array of generated auxiliary file names
}
\description{
Auxiliary files and metadata are appended to all data packages, but may also be
generated separately. Variable metadata is made available in tabular (.csv)
and in JSON (.json) format in conformity with Tabular Data Package RFC
specifications at \url{http://dataprotocols.org/tabular-data-package/}.
}
\details{
API call: Generate auxiliary files for soil organic carbon concentration and
2012 total population

\code{$ curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/datapackage \
 -d '{"var" : ["soc_d5", "PN12_TOT"]}' \
 -X POST -H 'Content-Type:application/json'
}
}
\examples{
# Generate auxiliary files for soil organic carbon concentration and 2012 total population
x <- datapackage(c("soc_d5", "PN12_TOT"))
x <- fread(x[1])
x

}
\seealso{
\link{indicator} and \link{category} for other convenience
methods to query HarvestChoice metadata.
}
