% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Return HarvestChoice variable metadata}
\usage{
describe(var, cat, version, raster = FALSE, as.class = "data.table",
  css = "json")
}
\arguments{
\item{var}{optional character array of HarvestChoice variable codes}

\item{cat}{optional category filter (fuzzy matches)}

\item{version}{optional version filter}

\item{raster}{return only raster variables}

\item{as.class}{simple "data.table" or if "list" group variables by category}

\item{css}{include Carto CSS rules}
}
\value{
a data.table or hierarchical list of variable metadata
}
\description{
Return HarvestChoice variable metadata
}
\seealso{
\code{\link{category}} to view HarvestChoice indicator categories
}

