% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominant.R
\name{dominant}
\alias{dominant}
\title{Return the dominant class (the mode of a classified variable)}
\usage{
dominant(x)
}
\arguments{
\item{x}{character array, can also be a factor}
}
\value{
a 1-length character of the dominant class
}
\description{
This is a helper method used to return the dominant class of a classified variable.
This is equivalent to median for a numeric variable.
}
\examples{
dominant(sample(letters, 100, replace=T))
}

