% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{g0}
\alias{g0}
\title{Country Boundaries}
\format{A SpatialPolygonsDataFrame with 56 obs. and 12 variables}
\usage{
data(g0)
}
\description{
GAUL 2008 country boundaries for sub-Saharan Africa used in HarvestChoice data layers.
}
\keyword{datasets}

