% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{g1}
\alias{g1}
\title{Province Boundaries}
\format{A SpatialPolygonsDataFrame with 3,240 obs. and 15 variables}
\usage{
data(g1)
}
\description{
GAUL 2008 province boundaries for sub-Saharan Africa used in HarvestChoice data layers.
}
\keyword{datasets}

