% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indicator.R
\name{genCartoCSS}
\alias{genCartoCSS}
\title{Generate CartoCSS snippet to symbolize raster tiles}
\usage{
genCartoCSS(var, pal = "BuGn", legend = TRUE, ...)
}
\arguments{
\item{var}{HarvestChoice variable code}

\item{pal}{optional Brewer color palette name used for plotting}

\item{legend}{if TRUE returns HarvestChoice legend, otherwise returns default legend}

\item{...}{any argument passed to \code{\link{getLayer}}, e.g. \code{by="FS_2012", iso3="GHA"}}

\item{iso3}{optional country or region filter (3-letter code)}
}
\value{
character, CartoCSS snippet for variable \code{var}
}
\description{
Generate CartoCSS snippet to symbolize raster tiles
}
\examples{
genCartoCSS("soc_d5")
}

