% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genCitation}
\alias{genCitation}
\title{Generate HarvestChoice data use terms, citation, and variable metadata}
\usage{
genCitation(var)
}
\arguments{
\item{var}{character array of CELL5M variable names}
}
\value{
character, path to generated README file
}
\description{
Generate HarvestChoice data use terms, citation, and variable metadata
}

