% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genFile}
\alias{genFile}
\title{Convert CELL5M layers to multiple formats and package for download}
\usage{
genFile(var, iso3 = "SSA", by = NULL, format = c("csv", "geojson", "tif",
  "shp", "dta", "asc", "rds", "rdata"), ...)
}
\arguments{
\item{format}{output format c("csv", "geojson", "tif", "shp", "dta", "asc", "rds", "rdata")}

\item{...}{any argument passed to getLayer(), e.g. \code{var}, \code{iso3}, \code{by}, and \code{geom}}
}
\value{
character, path to generated data file
}
\description{
Package any result from \code{getLayer()} into the user-specified data format.
Also creates an accompanying README file with metadata and citation details.
}

