% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getLayerWKT}
\alias{getLayerWKT}
\title{Summarize HarvestChoice 5-arc-minute layers over WKT geometri(es)}
\usage{
getLayerWKT(var, wkt, ...)
}
\arguments{
\item{var}{character array of variable names (all types are accepted)}

\item{wkt}{WKT representation of a spatial object (points, multipoints, or polygons)}

\item{...}{other arguments passed to \code{getLayer()} (by, iso3, collapse)}
}
\value{
a data.table of \code{var} indicators summarized across \code{wkt} geometry
}
\description{
This function intersects any WKT with HarvestChoice 5-arc-minute grid and returns
a list of intersected gridcell IDs (CELL5M codes) typically to pass on as an argument
to \code{getLayer()}.
}

