% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMeta}
\alias{getMeta}
\title{Return HarvestChoice variable metadata}
\usage{
getMeta(var, group, version, raster = FALSE, by.group = FALSE,
  css = "json")
}
\arguments{
\item{var}{character array of HarvestChoice variables}

\item{group}{optional category filter}

\item{version}{optional version filter}

\item{raster}{return only raster variables}

\item{by.group}{group variables by category}

\item{css}{include Carto CSS rules}
}
\value{
a data.table of variable metadata
}
\description{
Return HarvestChoice variable metadata
}

