% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPixelID}
\alias{getPixelID}
\title{Convert WKT objects to gridcell IDs}
\usage{
getPixelID(wkt)
}
\arguments{
\item{wkt}{WKT representation of a spatial object (points, multipoints, or polygons)}
}
\value{
integer array of gridcell IDs (CELL5M codes)
}
\description{
This function intersects any WKT with HarvestChoice 5-arc-minute grid and returns
a list of intersected gridcell IDs (CELL5M codes) typically to pass on as an argument
to \code{getLayer()}, e.g. \code{getLayer(var="whea_h", limit=getPixelID(""))}
}

