% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{getPlot}
\alias{getPlot}
\title{Generate single raster plot of HarvestChoice layer}
\usage{
getPlot(var, iso3 = "SSA", pal, format = "default", legend = "default",
  ...)
}
\arguments{
\item{var}{variable to plot (character of length 1)}

\item{iso3}{optional array of ISO3 country or region code to filter by}

\item{pal}{optional Brewer color palette used for plotting}

\item{format}{one of c("default", "print", "thumbnail") control legend and axes}

\item{legend}{one of c("default", "auto") uses HarvestChoice legend breaks or R default}

\item{...}{any argument passed to getLayer()}
}
\value{
plot
}
\description{
Generate single raster plot of HarvestChoice layer
}

