% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcapi.R
\name{hcapi}
\alias{hcapi}
\title{Query, subset, summarize, and download HarvestChoice indicators}
\usage{
hcapi(var, iso3 = "SSA", by = NULL, wkt = NULL, format = "data.table",
  ...)
}
\arguments{
\item{var}{character array of variable codes, passed to \code{\link{getLayer}}}

\item{iso3}{character array of ISO3 country or region codes, passed to \code{\link{getLayer}}}

\item{by}{character array of variable codes to summarize by, passed to \code{\link{getLayer}}}

\item{wkt}{WKT representation of a spatial object (points, multipoints, or polygons, multipolygons)}

\item{format}{output format, one of "data.table", "csv", "tif", "dta", "asc", "grd", "rds",
else "png" to plot the rasters, or "hist" to plot histogram and univariate statistics}

\item{...}{other optional arguments passed to \code{\link{getLayer}}, \code{\link{genFile}},
or to \code{\link{genPlot}}, e.g. \code{collapse}, \code{as.class}, \code{dir}, \code{pal}.}
}
\value{
a data.table (or other formats) of \code{var} indicators summarized by \code{by} domains
}
\description{
Wrapper method to query, subset and/or aggregate HarvestChoice layers.
This method may also be used to summarize classified variables along continuous
variables, e.g. \code{hcapi(var="AEZ16_CLAS", by="bmi")}. Here \code{AEZ16_CLAS} is
a classified (categorical) raster, and \code{bmi} is a continuous raster, but the
request is valid. The dominant class of \code{AEZ16_CLAS} is returned along intervals
of \code{bmi}. Default interval breaks are used but custom intervals may also be
defined, e.g. \code{hcapi(var="AEZ16_CLAS", by=list(bmi=c(0,5,10,15,20,25)))}.
The dominant class of a variable \code{var} is defined by \code{\link{dominant(var)}}.
Layers may also be summarized over spatial points are areas (passed as WKT representations
using argument \code{wkt}). Use the \code{format} argument to control the output
format (see examples below).
}
\seealso{
\code{\link{getLayer}}, \code{\link{getLayerWKT}}, \code{\link{genFile}}
}

