\name{hcapi3-package}
\alias{hcapi3-package}
\alias{hcapi3}
\docType{package}
\title{
HarvestChoice Data Query Services
}
\description{
Runs on Darby under FastRWeb root at
/var/FastRWeb/web.R/v2.0/methods.R Methods: - getLayer() -
makeZip() - genPlot() - genStats() - getMeta() -
genCartoCSS()
}
\details{
\tabular{ll}{
Package: \tab hcapi3\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-25\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Melanie Bacou <mel@mbacou.com>
Maintainer: Bacou, Melanie <mel@mbacou.com>
}
\references{
See HarvestChoice at \url{http://harvestchoice.org/}
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
