% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{indicator}
\alias{indicator}
\title{Return HarvestChoice indicator metadata}
\usage{
indicator(q, version = NULL, as.class = "data.table", cartoCSS = FALSE)
}
\arguments{
\item{q}{search string to match against HarvestChoice indicator codes,
labels, and categories. Will attempt a fuzzy match against these metadata elements.}

\item{version}{optional version filter}

\item{as.class}{"data.table" (default) or "list" for a list of indicators grouped
by category}

\item{cartoCSS}{also include formatted CartoCSS rules}
}
\value{
a data.table or hierarchical list of indicator metadata
}
\description{
Return HarvestChoice indicator metadata
}
\seealso{
\code{\link{category}} to view HarvestChoice indicator categories
}

