% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso}
\alias{iso}
\title{HarvestChoice Lookup Lists}
\format{arrays}
\usage{
iso.R
}
\description{
List of ISO3 country codes and corresponding country/region labels.
}
\examples{
# Lookup lists
names(iso)
}
\keyword{datasets}

