% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta}
\alias{meta}
\title{Write HarvestChoice CELL5M auxiliary files to CSV and JSON per Tabular Data
Package RFC specifications}
\usage{
meta(var, format = NULL, dir = ".")
}
\arguments{
\item{var}{character array of CELL5M variable codes}

\item{format}{character, file extension corresponding to the data package output
format. See \code{\link{genFile}}. If \code{NULL} no auxiliary file is produced
aside from metadata records in \code{meta.csv}.}

\item{dir}{character, optional output directory. Default to current working directory.}
}
\value{
character, paths to all generated auxiliary files
}
\description{
This information is appended to all data downloads, but may also be queried
separately. Variable metadata is made available in tabular format (.csv)
and in JSON (.json) format in conformity with the latest Tabular Data Package RFC
specifications documented at \url{http://dataprotocols.org/tabular-data-package/}.
}

