% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rankSimilar}
\alias{rankSimilar}
\title{Rank regions, districts, or pixels by similarity to a reference unit for a set of indicators}
\usage{
rankSimilar(x, var, by = 0, iso3 = "SSA")
}
\arguments{
\item{x}{region (ADM1_CODE), district (ADM2_CODE), or pixel code (CELL5M)}

\item{var}{character array of HarvestChoice variable codes}

\item{iso3}{optional country or region filter (3-letter code)}

\item{level}{reference geography, 1 for regions, 2 for districts, 0 for pixel)}
}
\value{
a data.table of regions, districts, or pixels ranked by similarity to \code{x}
}
\description{
Rank regions, districts, or pixels by similarity to a reference unit for a set of indicators
}

