% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme.R
\name{readme}
\alias{readme}
\title{Write HarvestChoice variable metadata to CSV}
\usage{
readme(var, file = "./readme.csv")
}
\arguments{
\item{var}{character array of CELL5M variable codes}

\item{file}{output file path, default ./README.csv in current directory}
}
\value{
character, path to generated README file
}
\description{
This information is appended to all data downloads, but may also be queried
separately. Variable metadata is now made available in tabular format (csv).
}

