% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Generate histogram and boxplot of HarvestChoice indicators}
\usage{
stats(var, iso3 = "SSA", by = NULL)
}
\arguments{
\item{var}{character array of variable codes to plot}

\item{iso3}{optional country or region filter (3-letter code)}
}
\value{
array of file names, one for each plot
}
\description{
Display a histogram and univariate statistics for any selected HarvestChoice variable(s).
See examples below. Note that calling \code{stats(...)} is equivalent to calling the
convenience method \code{hcapi(..., format="stats")}.
}

