% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vi}
\alias{vi}
\title{HarvestChoice CELL5M Variable Inventory}
\format{A data.table object}
\usage{
vi.R
}
\description{
Metadata elements for HarvestChoice collection of 5-arc-minute rasters for
sub-Saharan Africa.
}
\examples{
# List of metadata elements
names(vi)

# Full metadata for cassava yield
vi["cass_y"]
}
\keyword{datasets}

