% Generated by roxygen2 (4.0.2): do not edit by hand
\name{writePlot}
\alias{writePlot}
\title{Write layer plot to PNG}
\usage{
writePlot(var, iso3 = "SSA", format = "default", width = switch(format,
  640, print = 5, thumbnail = 180), height = switch(format, 640, print = 5,
  thumbnail = 200), units = switch(format, "px", print = "in", thumbnail =
  "px"), res = switch(units, `in` = 200, px = 72), cache = TRUE, ...)
}
\arguments{
\item{var}{character array of variables to plot}

\item{iso3}{optional country or region filter (3-letter code)}

\item{format}{format argument passed to genPlot()}

\item{width}{plot width in units passed to png()}

\item{height}{plot height in units passed to png()}

\item{units}{one of c("px", "in") passed to png()}

\item{res}{pixel-per-inch passed to png(), used if units="in"}

\item{cache}{if TRUE use image in server cache, if FALSE rebuild image}

\item{...}{parameters passed on to genPlot()}
}
\value{
character, path to generated ZIP file
}
\description{
Write layer plot to PNG
}

