% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/describe.R
\name{category}
\alias{category}
\title{Show HarvestChoice variable categories (3-level deep)}
\usage{
category(cat, as.class = "data.table")
}
\arguments{
\item{cat}{optional fuzzy filter}

\item{as.class}{"data.table" simple data table or "list" grouped by category}
}
\value{
a data.table with \code{N} showing the number of indicators in each category,
or a list of variable categories
}
\description{
Show HarvestChoice variable categories (3-level deep)
}
\examples{
category("Demographics")
category("cassava", as.class="list")
}

