% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme.R
\name{category}
\alias{category}
\title{Show HarvestChoice variable categories (3-level deep)}
\usage{
category(cat, as.class = "data.table")
}
\arguments{
\item{cat}{optional fuzzy filter}

\item{as.class}{"data.table" simple data table or "list" grouped by category}
}
\value{
a data.table with \code{N} showing the number of indicators in each category,
or a list of variable categories
}
\description{
Show HarvestChoice variable categories (3-level deep)
}
\examples{
# List all HarvestChoice indicators matching category "Demographics"
category("demographics")

# List all HarvestChoice indicators matching "cassava", return as a hierarchical list
category("cassava", as.class="list")

# Equivalent request using cURL at the command line and passing well-formatted JSON
# objects
# curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/category/json \\
# -d '{"cat" : "demographics'} \\
# -X POST -H "Content-Type:application/json"

#' # curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/category/json \\
# -d '{"cat" :" cassava", "as.class" : "list"} \\
# -X POST -H "Content-Type:application/json"
}

