% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/describe.R
\name{category}
\alias{category}
\title{Return list of HarvestChoice variable categories}
\usage{
category(cat, as.class = "data.table")
}
\arguments{
\item{cat}{optional category filter}

\item{as.class}{"data.table" simple data table or "list" group by category}
}
\value{
a data.table or list of variable categories
}
\description{
Return list of HarvestChoice variable categories
}
\examples{
category("Demographics")
}

