% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{category}
\alias{category}
\title{Return HarvestChoice indicator categories (3-level deep)}
\usage{
category(q = NULL, as.class = "data.table")
}
\arguments{
\item{q}{optional fuzzy filter}

\item{as.class}{"data.table" or "list" of indicator codes grouped by category}
}
\value{
a data.table showing the number of indicators in each category,
or a list of indicators grouped by category
}
\description{
Return HarvestChoice indicator categories (3-level deep)
}
\examples{
# List all HarvestChoice indicators matching category 'demographics'
category("demographics")

# List all HarvestChoice indicators matching 'cassava', return a hierarchical list
category("cassava", as.class="list")

# Equivalent request using cURL at the command line and passing well-formatted JSON
# objects
# curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/category/json \\
# -d '{"q" : "demographics'} \\
# -X POST -H "Content-Type:application/json"

#' # curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/category/json \\
# -d '{"q" :" cassava", "as.class" : "list"} \\
# -X POST -H "Content-Type:application/json"
}
\seealso{
\code{\link{indicator}} to view complete indicator metadata records
}

