% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{category}
\alias{category}
\title{Return HarvestChoice indicator categories (3-level deep)}
\usage{
category(q = NULL, as.class = "data.table")
}
\arguments{
\item{q}{character array of pattern(s) to search for. If omitted will return all
available indicators.}

\item{as.class}{"data.table" or "list" of indicator codes grouped by category}
}
\value{
a data.table showing the number of indicators in each category,
or a list of indicators grouped by category
}
\description{
Return a compact list of indicator categories, codes and labels
}
\details{
\code{
# API call: list all HarvestChoice indicators matching category 'demographics'
curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/category/json \
 -d '{"q" : "demographics'} \
 -X POST -H "Content-Type:application/json"

# API call: list all HarvestChoice indicators matching 'cassava' in a hierarchical list
curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/category/json \
 -d '{"q" :" cassava", "as.class" : "list"} \
 -X POST -H "Content-Type:application/json"

# To return a complete list of published indicators omit 'q'
curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/category/json \
 -X POST -H "Content-Type:application/json"
}
}
\examples{
# List all HarvestChoice indicators matching category 'demographics'
category("demographics")

# List all HarvestChoice indicators matching 'cassava' in a hierarchical list
category("cassava", as.class="list")
}
\seealso{
\code{\link{indicator}} to view complete indicator metadata records
}

