% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapackage.R
\name{datapackage}
\alias{datapackage}
\title{Write HarvestChoice auxiliary files to CSV and JSON per Tabular Data Package RFC
specifications}
\usage{
datapackage(var, format = NULL, dir = ".")
}
\arguments{
\item{var}{character array of HarvestChoice indicator code(s)}

\item{format}{file extension corresponding to the data package output
format. See \code{\link{genFile}}. If \code{NULL} no auxiliary file is produced
aside from metadata records in \code{meta.csv}.}

\item{dir}{character, optional output directory. Default to current working directory.}
}
\value{
character array of generated auxiliary file names
}
\description{
This information is appended to all data downloads, but may also be queried
separately. Variable metadata is made available in tabular (.csv)
and in JSON (.json) format in conformity with Tabular Data Package RFC
specifications documented at \url{http://dataprotocols.org/tabular-data-package/}.
}
\details{
\code{
# API call: Generate auxiliary files for soil organic carbon concentration and
# 2012 total population
curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/datapackage \
 -d '{"var" : ["soc_d5", "PN12_TOT"]}' \
 -X POST -H 'Content-Type:application/json'
}
}
\examples{
# Generate auxiliary files for soil organic carbon concentration and 2012 total population
x <- datapackage(c("soc_d5", "PN12_TOT"))
x

}
\seealso{
\code{\link{indicator}} and \code{\link{category}} for other convenience
methods to query HarvestChoice metadata.
}

