% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Return HarvestChoice variable metadata}
\usage{
describe(var, cat, version, raster = FALSE, by.category = FALSE,
  css = "json")
}
\arguments{
\item{var}{character array of HarvestChoice variables}

\item{cat}{optional category filter}

\item{version}{optional version filter}

\item{raster}{return only raster variables}

\item{css}{include Carto CSS rules}

\item{by.group}{group variables by category}
}
\value{
a data.table of variable metadata
}
\description{
Return HarvestChoice variable metadata
}
\examples{
# Show complete metadata for all variables under 'Cassava' category
describe(cat="cassava", raster=T, by.category=T)

# Equivalent cUrl request at the command line
# curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/describe/json \\
# -d '{"cat" : "cassava", "raster" : true, "by.category" : true}' \\
# -X POST -H "Content-Type:application/json"
}

