% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Return HarvestChoice variable metadata}
\usage{
describe(var, cat, version, raster = FALSE, as.class = "data.table",
  css = "json")
}
\arguments{
\item{var}{optional character array of HarvestChoice variable codes}

\item{cat}{optional category filter (fuzzy matches)}

\item{version}{optional version filter}

\item{raster}{return only raster variables}

\item{as.class}{simple "data.table" or if "list" group variables by category}

\item{css}{include Carto CSS rules}
}
\value{
a data.table or hierarchical list of variable metadata
}
\description{
Return HarvestChoice variable metadata
}
\examples{
# Show complete metadata for all variables under 'Cassava' category
describe(cat="cassava", raster=T)

# Equivalent cURL request at the command line
# curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/describe/json \\
# -d '{"cat" : "cassava", "raster" : true, "as.class" : "list"}' \\
# -X POST -H "Content-Type:application/json"

describe(raster=T, cat="population", as.class="list")[1:2]

# Equivalent cURL request at the command line
# curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/describe/json \\
# -d '{"raster" : true, "cat" : "population", as.class" : "list"}' \\
# -X POST -H "Content-Type:application/json"
}
\seealso{
\code{\link{category}} to view HarvestChoice indicator categories
}

