% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genCartoCSS}
\alias{genCartoCSS}
\title{Generate CartoCSS snippet to symbolize raster tiles}
\usage{
genCartoCSS(var, pal = "BuGn", legend = TRUE, ...)
}
\arguments{
\item{var}{HarvestChoice variable name}

\item{pal}{optional Brewer color palette used for plotting}

\item{legend}{if TRUE returns HarvestChoice legend, otherwise returns default legend}

\item{...}{any argument passed to getLayer(), e.g. by=FS_2012, iso3=GHA}

\item{iso3}{optional country or region filter (3-letter code)}
}
\value{
character, CartoCSS snippet for requested variable
}
\description{
Generate CartoCSS snippet to symbolize raster tiles
}

