% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genFile.R
\name{genFile}
\alias{genFile}
\title{Convert CELL5M layers to raster and/or tabular formats for download}
\usage{
genFile(var, iso3 = "SSA", by = NULL, format = c("csv", "geojson", "tif",
  "dta", "asc", "rda", "grd"), ...)
}
\arguments{
\item{var}{character array of variable codes, passed to \code\link{getLayer}}

\item{iso3}{character array of ISO3 country or region codes, passed to \code\link{getLayer}}

\item{by}{character array of variable codes to summarize by, passed to \code\link{getLayer}}

\item{format}{output format c("csv", "json", "tif", "dta", "asc", "grd", "rda")}

\item{...}{any other optional argument \code\link{getLayer}, e.g. \code{by}, \code{collapse}.}
}
\value{
character, array of generated file names
}
\description{
Package any result from \code\link{getLayer} into the user-specified tabular or raster
format. Also includes a README file with metadata and citation details. Currently
supported export formats include CSV (csv), STATA (dta), GeoJSON (json), GeoTIFF
(tif), R raster (grd), R Data (rda), and ESRI ASCII raster (asc).
}
\examples{
# Total wheat harvested area across 16 agro-ecological zones in Nigeria and Ethiopia
# in STATA format
genFile("whea_h", iso3=c("NGA", "ETH"), by="AEZ16_CLAS", format="dta")

# Sorghum production returned in ESRI ASCII raster format
genFile("sorg_p", format="asc")

# Equivalent cUrl requests at the command line
# curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/genFile/zip \\
# -d '{"var" : "whea_h", "iso3" : ["NGA", "ETH"], "by" : "AEZ16_CLAS", "format" : "dta"}' \\
# -X POST -H "Content-Type:application/json"

# curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/genFile/zip \\
# -d '{"var" : "sorg_p", "format" : "asc"}' \\
# -X POST -H "Content-Type:application/json"
}

