% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genFile.R
\name{genFile}
\alias{genFile}
\title{Download HarvestChoice indicators in raster and/or tabular formats}
\usage{
genFile(var, iso3 = "SSA", by = NULL, format = c("csv", "geojson", "tif",
  "dta", "asc", "rds", "grd"), dir = ".", ...)
}
\arguments{
\item{var}{character array of indicator codes, passed to \code{\link{getLayer}}}

\item{iso3}{character array of ISO3 country or region codes, passed to \code{\link{getLayer}}}

\item{by}{character array of indicator codes to summarize by, passed to \code{\link{getLayer}}}

\item{format}{output format, one of "csv", "json", "tif", "dta", "asc", "grd", "rds".}

\item{dir}{output directory, default to current working directory}

\item{...}{any other optional argument passed to \code{\link{getLayer}},
e.g. \code{as.class}, \code{collapse}.}
}
\value{
character, array of generated file names included in the data package
}
\description{
Package result from \code{\link{getLayer}} into any user-specified tabular or
spatial raster format. Also includes README and Tabular Data Package specifications.
Currently supported export formats include CSV (csv), STATA (dta), GeoJSON (geojson),
GeoTIFF (tif), R raster (grd), RData (rda), ESRI ASCII raster (asc), and netCDF (nc).
Calling \code{genFile(var="bmi", iso3="TZA", format="dta")} is equivalent to calling
the convenience method \code{hcapi(var="bmi", iso3="TZA", format="dta")}.
}
\seealso{
\code{\link{datapackage}} to generate associated metadata records
}

