% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genFile.R
\name{genFile}
\alias{genFile}
\title{Convert CELL5M layers to raster and/or tabular formats and create archive for download}
\usage{
genFile(var, iso3 = "SSA", by = NULL, format = c("csv", "geojson", "tif",
  "dta", "asc", "rda", "grd"), ...)
}
\arguments{
\item{var}{character array of variable codes, passed to getLayer(var, ...)}

\item{iso3}{character array of ISO3 country or region codes, passed to getLayer()}

\item{by}{character array of variable codes to summarize by, passed to getLayer()}

\item{format}{output format c("csv", "geojson", "tif", "dta", "asc", "grd", "rda")}

\item{...}{any other optional argument passed to getLayer(), e.g. \code{geom}, \code{collapse}.}
}
\value{
character, path to generated ZIP file
}
\description{
Package any result from \code{getLayer()} into the user-specified tabular or raster
format. Also includes a README file with metadata and citation details.
}
\examples{
# Total wheat harvested area across 16 agro-ecological zones in Nigeria and Ethiopia
# in STATA format
genFile("whea_h", iso3=c("NGA", "ETH"), by="AEZ16_CLAS", format="dta")

# Sorghum production returned in ESRI ASCII raster format
genFile("sorg_p", format="asc")

# Equivalent cUrl requests at the command line
# curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/genFile \\
# -d '{"var" : "whea_h", "iso3" : ["NGA", "ETH"], "by" : "AEZ16_CLAS", "format" : "dta"}' \\
# -X POST -H "Content-Type:application/json"

# Download generated ZIP file
# wget http://hcapi.harvestchoice.org/ocpu/tmp/x0fe45cc2aa/files/whea_h-AEZ16_CLAS-NGA.dta.zip

# curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/genFile \\
# -d '{"var" : "sorg_p", "format" : "asc"}' \\
# -X POST -H "Content-Type:application/json"

# Download generated ZIP file
# wget http://hcapi.harvestchoice.org/ocpu/tmp/x0760ca93d2/files/sorg_p--SSA.asc.zip
}

