% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genFile}
\alias{genFile}
\title{Convert CELL5M layers to multiple formats and package for download}
\usage{
genFile(var, iso3 = "SSA", by = NULL, format = c("csv", "geojson", "tif",
  "shp", "dta", "asc", "rds", "rdata"), ...)
}
\arguments{
\item{var}{character array of CELL5M variable names}

\item{iso3}{optional country or regional filter (3-letter code)}

\item{by}{optional character array of CELL5M variable to group by}

\item{format}{output format c("csv", "geojson", "tif", "dta", "asc", "rds", "RData")}

\item{...}{any argument passed to getLayer()}
}
\value{
character, path to generated data file
}
\description{
Convert CELL5M layers to multiple formats and package for download
}

