% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genPlot.R
\name{genPlot}
\alias{genPlot}
\title{Generate Raster Plot of HarvestChoice 5-Arc-Minute Spatial Indicators}
\usage{
genPlot(var, iso3 = "SSA", pal, format = "default", style = "default",
  width = 640, height = 640, ...)
}
\arguments{
\item{var}{character array of variable codes to plot}

\item{iso3}{optional ISO3 country or region code to filter by}

\item{pal}{optional Brewer color palette used for plotting, e.g. "Blues"}

\item{format}{one of "default", "print", "thumbnail" to control legend and axes}

\item{style}{one of \code{classIntervals()} style arguments (e.g. "kmeans"), or "default" to use default breaks}

\item{width}{plot width in pixel (unless \code{units} is specified)}

\item{height}{plot height in pixel (unless \code{units} is specified)}

\item{...}{any argument passed to \code{png()}, e.g. units, res, pointsize}
}
\value{
Array of generated file names, one for each plot
}
\description{
Generate Raster Plot of HarvestChoice 5-Arc-Minute Spatial Indicators
}
\examples{
# Generate standard raster plot of 2012 population density for sub-Saharan Africa
genPlot("PD12_TOT")

# Generate 3 raster plots for Ghana with legend and title but not axes
genPlot(c("AEZ16_CLAS", "whea_h", "soc_d15"), iso3="GHA", format="print")

# Generate 3 raster plots for Nigeria with the specified dimensions
p <- genPlot(c("FS_2012", "yield_l_cv", "soc_d15"), iso3="NGA", width=5, height=5,
units="in", res=200, pointsize=8)

# Print these plots
for (i in p) grid::grid.raster(png::readPNG(i))
}

