% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genPlot}
\alias{genPlot}
\title{Generate single raster plot of HarvestChoice layer}
\usage{
genPlot(var, iso3 = "SSA", pal, format = "default", legend = "default",
  ...)
}
\arguments{
\item{var}{variable to plot (character of length 1)}

\item{iso3}{optional ISO3 country code to filter on}

\item{pal}{optional Brewer color palette used for plotting}

\item{format}{one of c("default", "print", "thumbnail") control legend and axes}

\item{legend}{one of c("default", "auto") uses HarvestChoice legend breaks or R default}
}
\value{
plot
}
\description{
Generate single raster plot of HarvestChoice layer
}

