% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPlot.R
\name{genPlot}
\alias{genPlot}
\title{Plot HarvestChoice 5-arc-minute spatial indicators}
\usage{
genPlot(var, iso3 = "SSA", pal = character(0), layout = "default",
  style = NULL, ...)
}
\arguments{
\item{var}{character array of variable codes to plot}

\item{iso3}{optional ISO3 country or region code(s)}

\item{pal}{optional Brewer color palette used for plotting, e.g. "Blues"}

\item{layout}{one of "default" or "thumbnail" to control legend and axes}

\item{style}{passed to \code{\link[classInt:classIntervals]{classIntervals}}}

\item{...}{any argument passed to \code{\link[classInt:classIntervals]{classIntervals}},
e.g. \code{n=9} and \code{dataPrecision=0}.}
}
\value{
raster plot
}
\description{
Method to plot HarvestChoice rasters with mutiple layout and symbology options.
See examples below. Note that calling \code{genPlot(...)} is equivalent to calling
the convenience function \code{\link{hcapi}(..., format="plot")}.
}
\details{
API call: generate 2 plots showing farming systems and 2012 population density for Ghana

\code{$ curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/hcapi \
 -d '{"var":["FS_2012_TX", "PD12_TOT"], "iso3":"GHA", "format":"plot"}' \
 -X POST -H 'Content-Type:application/json'
--
/ocpu/tmp/x03d5aa8e98/R/.val
/ocpu/tmp/x03d5aa8e98/stdout
/ocpu/tmp/x03d5aa8e98/source
/ocpu/tmp/x03d5aa8e98/console
/ocpu/tmp/x03d5aa8e98/info
/ocpu/tmp/x03d5aa8e98/files/DESCRIPTION
/ocpu/tmp/x03d5aa8e98/graphics/1
/ocpu/tmp/x03d5aa8e98/graphics/2
}

GET all generated plots in a ZIP archive

\code{
$ wget http://hcapi.harvestchoice.org/ocpu/tmp/x03d5aa8e98/graphics/last/png?width=640
$ wget http://hcapi.harvestchoice.org/ocpu/tmp/x03d5aa8e98/graphics/last/svg
}
}
\examples{
# Generate standard raster plot of 2012 population density for sub-Saharan Africa
genPlot("PD12_TOT", pal="YlOrRd", style="pretty", n=8)

# Generate 3 raster plots for Ghana with legend and title but not axes
genPlot(c("AEZ16_CLAS", "cass_h"), iso3="GHA", layout="thumbnail")

# Generate 3 raster plots for Nigeria with the specified dimensions
hcapi(c("FS_2012", "yield_l_cv", "soc_d15"), iso3="NGA", format="plot")

}
\seealso{
\link{hcapi} link\[classInt:classIntervals]{classIntervals}
}
