% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPlot.R
\name{genPlot}
\alias{genPlot}
\title{Plot HarvestChoice 5-arc-minute spatial indicators}
\usage{
genPlot(var, iso3 = "SSA", pal = character(0), layout = "default",
  style = "default", n = integer(0), width = switch(layout, default = 640,
  print = 5, thumbnail = 120), height = switch(layout, default = 640, print =
  5, thumbnail = 120), units = switch(layout, default = "px", print = "in",
  thumbnail = "px"), res = switch(layout, default = NA, print = 300, thumbnail
  = NA), ...)
}
\arguments{
\item{var}{character array of variable codes to plot}

\item{iso3}{optional ISO3 country or region code to filter by}

\item{pal}{optional Brewer color palette used for plotting, e.g. "Blues"}

\item{layout}{one of "default", "print", or "thumbnail" to control legend and axes}

\item{style}{one of \code{\link[classInt:classIntervals]{classIntervals}} \code{style}}

\item{n}{\code{\link[classInt:classIntervals]{classIntervals}} \code{n} argument
to control the number of breaks}

\item{width}{plot width in pixel (unless \code{units} is specified)}

\item{height}{plot height in pixel (unless \code{units} is specified)}

\item{units}{one of "px" (default), "in", "cm" or "mm".
Passed to \code{\link[grDevices:png]{png}}}

\item{...}{any argument passed to \code{\link[grDevices:png]{png}}, e.g. pointsize}

\item{resolution}{in ppi, by default set to 300ppi for print layout.
Passed to \code{\link[grDevices:png]{png}}
options (e.g. "kmeans" or "pretty") or "default" to use default breaks}
}
\value{
Array of generated file names, one for each plot
}
\description{
Method to plot HarvestChoice rasters with mutiple layout and symbology options.
See examples below. Note that calling \code{genPlot(...)} is equivalent to calling
the convenience function \code{hcapi(..., format="png")}.
}

