% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genRank.R
\name{genRank}
\alias{genRank}
\title{Rank agricultural productions across geographies}
\usage{
genRank(var = "h", by = "ISO3", ...)
}
\arguments{
\item{var}{one or several of "h", "p", or "v" for ranking by harvested area (h),
production (p), and/or value of production (v).
This argument may be expanded in the future to include additional sub-groups of indicators.}

\item{by}{by default returns aggregate production across all selected countries.
\code{by} may be specified to return aggregates across sub-zones (e.g. across provinces,
districts, agro-ecological zones, etc.), see examples.}

\item{...}{optional arguments passed to \code{\link{hcapi}}, e.g. \code{iso3} and \code{wkt}}
}
\value{
data.table(s) of commodities ranked from top to bottom
}
\description{
Returns top and bottom commodities in terms of aggregate production, harvested area,
and/or value of production.
}
\seealso{
\link{hcapi}
}

