% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genRank.R
\name{genRank}
\alias{genRank}
\title{Rank agricultural productions across geographies}
\usage{
genRank(var = "h", by = "ISO3", ...)
}
\arguments{
\item{var}{one or several of "harvested area", "production", or "value of production".
These argument can be abbreviated to "h", "p" or "v". May be expanded in the future
to include additional sub-groups of indicators.}

\item{by}{by default returns aggregate production across all selected countries.
\code{by} may be specified to return aggregates across sub-zones (e.g. across provinces,
districts, agro-ecological zones, etc.), see examples.}

\item{...}{optional arguments passed to \code{\link{hcapi}}, e.g. \code{iso3} and \code{wkt}}
}
\value{
data.table(s) of commodities ranked from top to bottom
}
\description{
Returns top and bottom commodities in terms of aggregate production, harvested area,
and/or value of production.
}
\details{
API call: rank commodities in Angola and Mozambique by harvested area and production

\code{$ curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/genRank/json \
 -d '{"var" : ["h", "p"], "iso3" : ["AGO", "MOZ"]}' \
 -X POST -H 'Content-Type:application/json'
}

API call: rank commodities across Angola's agro-ecological zones by production

\code{$ curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/genRank/json \
 -d '{"var" : "p", "iso3" : "AGO", "by" : "AEZ5_CLAS"}' \
 -X POST -H 'Content-Type:application/json'
}
}
\examples{
# Rank commodities in Angola and Mozambique by harvested area and production
x <- genRank(c("h", "p"), iso3=c("AGO", "MOZ"))
x

# Rank commodities across Angola's agro-ecological zones by production
x <- genRank("p", iso3="AGO", by="AEZ5_CLAS")
x

}
\seealso{
\link{hcapi}
}
