% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meta.R
\name{genReadme}
\alias{genReadme}
\title{Generate HarvestChoice data use terms, citation, and variable metadata}
\usage{
genReadme(var)
}
\arguments{
\item{var}{character array of CELL5M variable names}
}
\value{
character, path to generated README file
}
\description{
Generate HarvestChoice data use terms, citation, and variable metadata
}

