% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{genReadme}
\alias{genReadme}
\title{Return HarvestChoice data use terms, citation, and variable metadata}
\usage{
genReadme(var)
}
\arguments{
\item{var}{character array of CELL5M variable codes}
}
\value{
character, path to generated README file
}
\description{
Return HarvestChoice data use terms, citation, and variable metadata
}
\examples{
# README for soil organic carbon concentration
f <- genReadme("soc_d5")
print(readLines(f))

}

