% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/describe.R
\name{genReadme}
\alias{genReadme}
\title{Generate HarvestChoice data use terms, citation, and variable metadata}
\usage{
genReadme(var)
}
\arguments{
\item{var}{character array of CELL5M variable names}
}
\value{
character, path to generated README file
}
\description{
Generate HarvestChoice data use terms, citation, and variable metadata
}
\examples{
# README for soil organic carbon concentration
f <- genReadme("sod_d5")
print(readLines(f))
}

