% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{genStats}
\alias{genStats}
\title{Generate histogram and boxplot of HarvestChoice layer(s)}
\usage{
genStats(var, iso3 = "SSA", by = NULL)
}
\arguments{
\item{var}{character array of variables to summarize}

\item{iso3}{optional country or region filter (3-letter code)}
}
\value{
plot
}
\description{
Generate histogram and boxplot of HarvestChoice layer(s)
}

