% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGroups}
\alias{getGroups}
\title{Return list of HarvestChoice variable categories}
\usage{
getGroups(group, format = "data.table")
}
\arguments{
\item{group}{optional category filter}

\item{format}{one of "data.table" (default), "json", or "html" for a human-readable table}
}
\value{
a data.table of variable categories
}
\description{
Return list of HarvestChoice variable categories
}

