% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLayer.R
\name{getLayer}
\alias{getLayer}
\title{Subset, and/or summarize HarvestChoice 5-arc-minute spatial indicators}
\usage{
getLayer(var, iso3 = "SSA", by = NULL, ids = NULL, collapse = TRUE,
  as.class = "data.table")
}
\arguments{
\item{var}{character array of variable names (all types are accepted)}

\item{iso3}{optional array of country or regional code(s) to filter by (use 3-letter code).
Use \code{iso} to view all avalaible codes.}

\item{by}{optional character array of variables to group by (all types are accepted)}

\item{ids}{optional gridcell ids to filter by (if collapse=F) or summarize by (if collapse=T)}

\item{collapse}{if TRUE collapses over \code{by} variables. If FALSE always return
all pixel values (useful for plotting and to convert to spatial formats).}

\item{as.class}{one of "data.table" (default) or "list". By default returns a simple data.table.
If \code{as.class="list"} returns a well-constructed list with variable metadata.}
}
\value{
a data.table (or json array) of \code{var} indicators aggregated by \code{by} domains
}
\description{
Workhorse method to subset and/or aggregate HarvestChoice layers.
This method also aggregates classified variables by continuous variables.\\
e.g. \code{getLayer(var="AEZ16_CLAS", by="bmi")}.
It does so by returning the dominant class of a classified variable within each \code{by}
class, and by automatically classifying any continuous variable passed to \code{by}
using default value breaks as specified in the variable metadata.
The dominant class of a variable \code{x} is defined as \code{\link{dominant}(x)}.
Layers may also be summarized over a spatial area (passed as an integer array of CELL5M IDs).
Note that calling \code{getLayer(...)} is equivalent to using the convenience method
\code{hcapi(...)} with the same arguments.
}
\seealso{
\link{hcapi} \link{getLayerWKT}
}

