% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getLayer}
\alias{getLayer}
\title{Subset, and/or aggregate HarvestChoice 5-arc-minute layers}
\usage{
getLayer(var, iso3 = "SSA", by = NULL, collapse = TRUE)
}
\arguments{
\item{var}{character array of variable names}

\item{iso3}{optional country or regional filter (3-letter code)}

\item{by}{optional character array of variables to group by}

\item{collapse}{if FALSE always return all pixel values (useful for plotting)}
}
\value{
a data.table of \code{varCode} aggregated by \code{by}
}
\description{
Workhorse method to subset and/or aggregate HarvestChoice layers.
This method also aggregates classified variables by continuous variables.
}

