% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLayer.R
\name{getLayer}
\alias{getLayer}
\title{Query, subset, and/or summarize HarvestChoice indicators}
\usage{
getLayer(var, iso3 = "SSA", by = NULL, ids = NULL, collapse = TRUE,
  as.class = "data.table")
}
\arguments{
\item{var}{character array of variable names (all types are accepted). Use e.g.
\code{link{category("poverty")}} to search for valid variable codes).}

\item{iso3}{optional array of 3-letter country or regional code(s) to filter by.
Use \code{iso} to view all available codes.}

\item{by}{optional character array of variables to summarize by (all types are accepted)}

\item{ids}{optional array of gridcell ids to filter by (if collapse=FALSE) or
summarize by (if collapse=TRUE).}

\item{collapse}{if TRUE collapses over \code{by} variables. If FALSE always return
all pixel values (useful for plotting and to convert to spatial formats).}

\item{as.class}{one of "data.table" (default) or "list". By default returns a simple
data.table. If \code{as.class="list"} returns a list with variable metadata.}
}
\value{
a data.table (or json array) of \code{var} indicators aggregated by
\code{by} domains
}
\description{
Workhorse method to query, subset and/or aggregate HarvestChoice layers.
This method may also be used to summarize classified variables along continuous
variables, e.g. \code{getLayer(var="AEZ16_CLAS", by="bmi")}.
It does so by returning the dominant class of a classified variable within each \code{by}
class, and by splitting any continuous variable passed to \code{by} using default value
breaks. The dominant class of a variable \code{var} is defined as \code{\link{dominant}(var)}.
Variables may also be summarized over custom areas (passed as an integer array of
gridcell IDs). Note that calling \code{getLayer(...)} is equivalent to using the
convenience method \code{hcapi(...)} with the same arguments.
}
\seealso{
\link{hcapi} and \link{getLayerWKT}
}

