% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLayerWKT.R
\name{getLayerWKT}
\alias{getLayerWKT}
\title{Summarize HarvestChoice 5-arc-minute layers over WKT geometri(es)}
\usage{
getLayerWKT(var, wkt, ...)
}
\arguments{
\item{var}{character array of variable names (all types are accepted)}

\item{wkt}{WKT representation of a spatial object (points, multipoints, or polygons)}

\item{...}{other arguments passed to \code\link{getLayer} (e.g. \code{by},
\code{iso3}, \code{collapse})}
}
\value{
a data.table of \code{var} indicators summarized across \code{wkt} geometry
}
\description{
This function intersects any spatial object (WKT point, multipoints, or polygons)
with HarvestChoice 5-arc-minute grid and returns a list of intersected gridcell IDs
(CELL5M codes) to pass along as argument to \code\link{getLayer}.
}

