% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcapi.R
\name{hcapi}
\alias{hcapi}
\title{Subset, summarize, and download HarvestChoice 5-arc-minute spatial indicators}
\usage{
hcapi(var, iso3 = "SSA", by = NULL, wkt = NULL, format = "default", ...)
}
\arguments{
\item{var}{character array of variable codes, passed to \code{\link{getLayer}}}

\item{iso3}{character array of ISO3 country or region codes, passed to \code{\link{getLayer}}}

\item{by}{character array of variable codes to summarize by, passed to \code{\link{getLayer}}}

\item{wkt}{WKT representation of a spatial object (points, multipoints, or polygons)}

\item{format}{output format, one of "default" (R data.table, default), "csv", "json", tif",
"dta", "asc", "grd", "rds", else "png" to plot the rasters, or "stats" to plot
histogram and univariate statistics.}

\item{...}{other optional arguments passed to \code{\link{getLayer}},
e.g. \code{collapse}, \code{as.class}}
}
\value{
a data.table (or other formats) of \code{var} indicators aggregated by \code{by} domains
}
\description{
Main method to subset and/or aggregate HarvestChoice layers.
This method also aggregates classified variables by continuous variables.\\
e.g. \code{hcapi(c("whea_h", "AEZ16_CLAS"), by=c("ADM2_NAME_ALT", "bmi"))}.
It does so by returning the dominant class of a classified variable within each \code{by}
class, and by automatically classifying any continuous variable passed to \code{by}
using default value breaks as specified in the variable metadata.
The dominant class of \code{x} is defined as \code{names(which.max(table(x)))}. For
convenience this function is wrapped as \code{\link{dominant}} in this package.
Layers can also be summarized over a spatial area (passed as an integer array of CELL5M ids).
Use \code{format} argument to control the output format for the spatial layer (see below).
}

