% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcapi.R
\name{hcapi}
\alias{hcapi}
\title{Subset, summarize, and download HarvestChoice 5-arc-minute spatial indicators}
\usage{
hcapi(var, iso3 = "SSA", by = NULL, wkt = NULL, format = NULL, ...)
}
\arguments{
\item{var}{character array of variable codes, passed to \code{\link{getLayer}}}

\item{iso3}{character array of ISO3 country or region codes, passed to \code{\link{getLayer}}}

\item{by}{character array of variable codes to summarize by, passed to \code{\link{getLayer}}}

\item{wkt}{WKT representation of a spatial object (points, multipoints, or polygons, multipolygons)}

\item{format}{output format, one of "csv", "json", tif", "dta", "asc", "grd", "rds",
else "png" to plot the rasters, or "stats" to plot histogram and univariate statistics}

\item{...}{other optional arguments passed to \code{\link{getLayer}} or \code{\link{genFile}},
e.g. \code{collapse}, \code{as.class}, \code{dir}}
}
\value{
a data.table (or other formats) of \code{var} indicators aggregated by \code{by} domains
}
\description{
Main method to subset and/or aggregate HarvestChoice layers.
This method also aggregates classified variables by continuous variables.\\
e.g. \code{hcapi(var="AEZ16_CLAS", by="bmi")}. Here \code{AEZ16_CLAS} is a classified
variable, and \code{bmi} is a continuous variable, but the request is valid.
The dominant class of \code{AEZ16_CLAS} is returned along intervals of \code{bmi}.
By default intervals are read from the variable metadata, but custom intervals may
also be defined, e.g. \code{hcapi(var="AEZ16_CLAS", by=list(bmi=c(0,5,10,15,20,25)))}.
The dominant class of a variable \code{x} is defined by \code{\link{dominant}}.
Layers may also be summarized over a spatial area (passed as an integer array of CELL5M IDs)
using parameter \code{ids}. Use \code{format} argument to control the output format
(see below).
}
\seealso{
\code{\link{getLayer}}, \code{\link{getLayerWKT}}, \code{\link{genFile}}
}

