% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indicator.R
\name{indicator}
\alias{indicator}
\title{Return HarvestChoice indicator metadata}
\usage{
indicator(q, version = NULL, as.class = "data.table", cartoCSS = FALSE)
}
\arguments{
\item{q}{search string to match against HarvestChoice indicator codes,
labels, and categories. Will attempt a fuzzy match against these metadata elements.}

\item{version}{optional version filter}

\item{as.class}{"data.table" (default) or "list" for a list of indicators grouped
by category}

\item{cartoCSS}{also include formatted CartoCSS rules}
}
\value{
a data.table or hierarchical list of indicator metadata
}
\description{
Return HarvestChoice indicator metadata
}
\examples{
# Show complete metadata for all HarvestChoice indicators matching 'cassava' or
'population'
indicator("cassava")

# Equivalent cURL request at the command line
# curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/indicator/json \\
# -d '{"q" : "cassava", "as.class" : "list"}' \\
# -X POST -H "Content-Type:application/json"

indicator("population", as.class="list")[1:2]

# Equivalent cURL request at the command line
# curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/indicator/json \\
# -d '{"q" : "population", as.class" : "list"}' \\
# -X POST -H "Content-Type:application/json"
}
\seealso{
\code{\link{category}} to view HarvestChoice indicator categories
}

