% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso}
\alias{iso}
\title{List of ISO3 country and/or region codes}
\format{arrays}
\usage{
iso.R
}
\description{
List of ISO3 country codes and corresponding country/region labels.
}
\note{
Country boundaries are derived from FAO GAUL 2008 (2009 eds.)
}
\examples{
# Lookup lists
iso

}
\keyword{datasets}
