% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{lookup}
\alias{lookup}
\title{HarvestChoice Lookup Lists}
\format{arrays}
\usage{
data(lookup)
}
\description{
Metadata for HarvestChoice collection of over 700 5-arc-minute rasters for sub-Saharan Africa.
}
\keyword{datasets}

