% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup}
\alias{lookup}
\title{HarvestChoice Lookup Lists}
\format{arrays}
\usage{
lookup.R
}
\description{
Metadata for HarvestChoice collection of over 700 5-arc-minute rasters for sub-Saharan Africa.
}
\examples{
# Lookup lists
names(iso)
}
\keyword{datasets}

