% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme.R
\name{readme}
\alias{readme}
\title{Return HarvestChoice data use terms, citation, and variable metadata}
\usage{
readme(var)
}
\arguments{
\item{var}{character array of CELL5M variable codes}
}
\value{
character, path to generated README.csv file
}
\description{
This information is appended to all data downloads, but may also be queried
separately. Variable metadata is now made available in tabular format (tsv).
}

