% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Generate histogram and boxplot of HarvestChoice indicators}
\usage{
stats(var, iso3 = "SSA", by = NULL)
}
\arguments{
\item{var}{character array of variable codes to plot}

\item{iso3}{optional country or region filter (3-letter code)}
}
\value{
array of file names, one for each plot
}
\description{
Display a histogram and univariate statistics for any selected HarvestChoice variable(s).
See examples below. Note that calling \code{stats(...)} is equivalent to calling the
convenience method \code{hcapi(..., format="stats")}.
}
\examples{
# Plots of BMI and cassava yield distribution over sub-Saharan Africa 10km grid
stats(c("bmi", "cass_y"))

#' # This method may be called via HTTP POST request using e.g. cUrl at the command line
# Return 2 plots showing farming systems and 2012 population density in Ghana
# curl http://hcapi.harvestchoice.org/ocpu/library/hcapi3/R/stats \\
# -d '{"var":["bmi", "cass_y"]}' \\
# -X POST -H 'Content-Type:application/json'

# /ocpu/tmp/x09db409895/R/.val
# /ocpu/tmp/x09db409895/graphics/1
# /ocpu/tmp/x09db409895/graphics/2
# /ocpu/tmp/x09db409895/source
# /ocpu/tmp/x09db409895/console
# /ocpu/tmp/x09db409895/info
# /ocpu/tmp/x09db409895/files/DESCRIPTION

# Use wget (at the command line) to download the generated plot(s)
# wget http://hcapi.harvestchoice.org/ocpu/tmp/x09db409895/graphics/1/png
# wget http://hcapi.harvestchoice.org/ocpu/tmp/x09db409895/graphics/2/png
}

