% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Generate histogram and boxplot of HarvestChoice layer(s)}
\usage{
stats(var, iso3 = "SSA", by = NULL)
}
\arguments{
\item{var}{character array of variables to summarize}

\item{iso3}{optional country or region filter (3-letter code)}
}
\value{
plot
}
\description{
Generate histogram and boxplot of HarvestChoice layer(s)
}
\examples{
# Plots of BMI and cassava yield distribution over sub-Saharan Africa 10km grid
stats(c("bmi", "cass_y"))
}

