% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vi}
\alias{vi}
\title{HarvestChoice CELL5M Variable Inventory}
\format{A data.table object}
\usage{
vi.R
}
\description{
Metadata for HarvestChoice collection of over 700 5-arc-minute rasters for sub-Saharan Africa.
}
\keyword{datasets}

