% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{vi}
\alias{vi}
\title{HarvestChoice CELL5M Variable Inventory}
\format{A data.table object}
\usage{
data(vi)
}
\description{
Metadata for HarvestChoice collection of over 700 5-arc-minute rasters for sub-Saharan Africa.
}
\keyword{datasets}

