% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vi}
\alias{vi}
\title{HarvestChoice inventory of 5-arc-minute spatial indicators}
\format{A data.table object}
\usage{
vi.R
}
\description{
Metadata elements for HarvestChoice collection of 5-arc-minute rasters for
sub-Saharan Africa.
}
\examples{
# List of metadata elements that may be accessed though `vi`
names(vi)

# Print full metadata for cassava yield
knitr::kable(data.frame(t(vi["cass_y"])), "rst", row.names=T)

# Get specific metadata elements
vi["cass_y"][, list(varTitle, units)]
}
\keyword{datasets}

