% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vi}
\alias{vi}
\title{Catalog of HarvestChoice indicators}
\format{A data.table object}
\usage{
vi.R
}
\description{
Complete metadata records for HarvestChoice 5-arc-minute inidcators.
}
\examples{
# List of metadata elements that may be accessed though `vi`
names(vi)

# Print full metadata for cassava yield
t(vi["cass_y"])

# Get specific metadata elements
vi["cass_y", .(varTitle, unit)]

# Count of layers by category
vi[, .N, by=.(cat1, cat2)]
}
\keyword{datasets}

