% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vi}
\alias{vi}
\title{HarvestChoice inventory of 5-arc-minute spatial indicators}
\format{A data.table object}
\usage{
vi.R
}
\description{
Metadata elements for HarvestChoice collection of 5-arc-minute rasters for
sub-Saharan Africa.
}
\examples{
# List of metadata elements that may be accessed though `vi`
names(vi)

# Print full metadata for cassava yield
x <- as.list(t(vi["cass_y"]))
names(x) <- names(vi)
x

# Get specific metadata elements
vi["cass_y"][, list(varTitle, units)]

# Count of layers by category
vi[, .N, by=list(cat1, cat2)]
}
\keyword{datasets}

